Change exit_code type from DWORD to int
- Updated exit_code variable from DWORD to int
- Adjusted CreateAndWaitForProcess and RunScript signatures to use int exit codes
- Ensured consistency with WinMain’s int return type

+++ src/error.h
@@ -1,8 +1,8 @@
 #include <stdarg.h>
 
 // Exit code definitions
-#define EXIT_CODE_SUCCESS ((DWORD)0)
-#define EXIT_CODE_FAILURE ((DWORD)-1)
+#define EXIT_CODE_SUCCESS (0)
+#define EXIT_CODE_FAILURE (-1)
 
 /**
  * EnableDebugMode - Enable debug mode
